pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pathology-llava"
version = "0.1.0"
description = "Hugging Face-compatible wrapper for PA-LLaVA / Pathology-LLaVA using PLIP and LLaMA-3."
authors = [
  { name = "Stefano Roy Bisignano" }
]
license = { text = "Apache-2.0" }
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "torch>=2.1",
  "transformers>=4.40.0",
  "huggingface_hub>=0.23.0",
  "Pillow>=10.0",
]

[project.urls]
Homepage = "https://github.com/<your-user>/pathology-llava"
Source = "https://github.com/<your-user>/pathology-llava"
Issues = "https://github.com/<your-user>/pathology-llava/issues"

[tool.setuptools.packages.find]
where = ["src"]
